requirements.txt
connexion==2.4.0
Flask==1.1.1
pytest==5.3.1
flask-cors==3.0.8
boto3==1.11.0
pyyaml==5.1.2
